Avoid pre-computation of possible search paths
Rather than enumerate the possibilities and walk them, prefer to
generate the possibilities as necessary.  This avoids the storage cost
for all possible paths.

Rename the interface to `ExecutableFinder` to avoid confusion over the
fact that this is not a generic file search utility.  Additionally,
rename the helper to `which(_:paths:extensions:)` as it correlates to
the Unix `which` command more closely (Windows' `where` will list all
entries).